feat: Add `-P` flag to print with newline
Before, `-p` would automatically print with newline. But this is
problematic for functions like `term.cursor_down`, since only the escape
without the newline would be emitted. This fixes that.

This also is a breaking change because it no longer allows passing
multiple arguments. That is, passing `-p -d` is no longer possible. It
must be passed like `-pd` now. This simplifies the processing of
arguments